pyproject.toml
[project]
name = "mcp-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.7.1",
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.3",
    "openai>=1.86.0",
    "python-dotenv>=1.1.0",
]
